FIX: bugs

+++ textbox/model/LM/gpt2.py
@@ -100,7 +100,7 @@ class GPT2(UnconditionalGenerator):
         loss = loss.reshape_as(decoder_labels)
 
         length = (decoder_labels != self.padding_token_idx).sum(dim=1).float()
-        loss = loss.sum(dim=1) / length
+        loss = loss.sum(dim=1) / length.float()
         return loss.mean()
 
     def calculate_nll_test(self, corpus, epoch_idx=-1):
